Added i18n hooks to plugins.
(This is a fix for Mantis bug 2000)

The PluginController now has additional methods to request the default language bundle, and also language bundles for different languages and countries.  These are then merged into the language bundle in use by Thinlet/UiGeneratorController, taking care not to overwrite values that are already present.