Updated Plugin management and features.
There is now a menu for enabling and disabling plugins.
The name and icon of a plugin can be specified using an annotation.
Running plugins are now kept track of by a PluginManager instance rather than the FrontlineSMS object as was previous.
Plugin spring and hibernate config are now loaded regardless of whether a plugin is activated, as they will be needed if it is activated later from the menu.