Fixed text getter bug in Thinlet.
The language resource bundle handling in thinlet had been modified, introducing a bug whereby the value from the default text resource would not be loaded if no value could be found in the standard text resource.  This has been fixed.

Also added some notes on how this should be fixed permanently, and other aspects of the language bundle handling.