Fixed bug with plugin text resource loading.
The bug was an omission in the loading of the text in UiGeneratorController.  When a plugin was loaded from the start of the application, the text resources would be loaded.  When the plugin was loaded later, the text resource loading code was never called.  This code has been added.

Also fixed a bug in the way non-default text bundles were being loaded for plugins - the LanguageBundle class previously provided the human-readable name of the language rather than the ISO code when creating the locale for requesting non-default text bundles.  This has been fixed.